+++ hub/src/server/errors.ts
@@ -49,3 +49,10 @@ export class InvalidInputError extends Error {
     this.name = this.constructor.name
   }
 }
+
+export class PayloadTooLargeError extends Error {
+  constructor (message: string) {
+    super(message)
+    this.name = this.constructor.name
+  }
+}