+++ hub/src/server/http.ts
@@ -68,13 +68,15 @@ export function makeHttpServer(config: HubConfigInterface): { app: express.Appli
         if (err instanceof errors.ValidationError) {
           writeResponse(res, { message: err.message, error: err.name }, 401)
         } else if (err instanceof errors.AuthTokenTimestampValidationError) {
-          writeResponse(res, { message: err.message, error: err.name  }, 401)
+          writeResponse(res, { message: err.message, error: err.name }, 401)
         } else if (err instanceof errors.BadPathError) {
-          writeResponse(res, { message: err.message, error: err.name  }, 403)
+          writeResponse(res, { message: err.message, error: err.name }, 403)
         } else if (err instanceof errors.NotEnoughProofError) {
-          writeResponse(res, { message: err.message, error: err.name  }, 402)
+          writeResponse(res, { message: err.message, error: err.name }, 402)
         } else if (err instanceof errors.ConflictError) {
-          writeResponse(res, { message: err.message, error: err.name  }, 409) 
+          writeResponse(res, { message: err.message, error: err.name }, 409) 
+        } else if (err instanceof errors.PayloadTooLargeError) {
+          writeResponse(res, { message: err.message, error: err.name }, 413)
         } else {
           writeResponse(res, { message: 'Server Error' }, 500)
         }
@@ -148,9 +150,9 @@ export function makeHttpServer(config: HubConfigInterface): { app: express.Appli
   })
 
   app.post(
-    /^\/list-files\/([a-zA-Z0-9]+)\/?/, express.json(),
+    /^\/list-files\/([a-zA-Z0-9]+)\/?/, express.json({ limit: 4096 }),
     (req: express.Request, res: express.Response) => {
-      // sanity check...
+      // sanity check... should never be reached if the express json parser is working correctly
       if (parseInt(req.headers['content-length']) > 4096) {
         writeResponse(res, { message: 'Invalid JSON: too long'}, 400)
         return
@@ -179,9 +181,9 @@ export function makeHttpServer(config: HubConfigInterface): { app: express.Appli
 
   app.post(
     /^\/revoke-all\/([a-zA-Z0-9]+)\/?/, 
-    express.json(),
+    express.json({ limit: 4096 }),
     (req: express.Request, res: express.Response) => {
-      // sanity check...
+      // sanity check... should never be reached if the express json parser is working correctly
       if (parseInt(req.headers['content-length']) > 4096) {
         writeResponse(res, { message: 'Invalid JSON: too long'}, 400)
         return
@@ -225,6 +227,7 @@ export function makeHttpServer(config: HubConfigInterface): { app: express.Appli
     const readURLPrefix = server.getReadURLPrefix()
     writeResponse(res, { 'challenge_text': challengeText,
                          'latest_auth_version': LATEST_AUTH_VERSION,
+                         'max_file_upload_size_megabytes': server.maxFileUploadSizeMB,
                          'read_url_prefix': readURLPrefix }, 200)
   })